Expose conversion code as package (#1)
