+++ .github/workflows/go.yml
@@ -17,12 +17,12 @@ jobs:
       uses: actions/setup-go@v2
       with:
         go-version: '^1.15'
+    - name: Checkout code
+      uses: actions/checkout@v2
     - name: Install protoc
       uses: arduino/setup-protoc@v1
     - name: Install protoc-gen-go
       run: go get -u github.com/golang/protobuf/protoc-gen-go
-    - name: Checkout code
-      uses: actions/checkout@v2
 
     # Formatting, go mod tidy, and re-generate proto extension code
     - name: Run go fmt on all modules
