+++ go.mod
@@ -5,5 +5,5 @@ go 1.15
 require (
 	github.com/golang/glog v1.0.0
 	github.com/golang/protobuf v1.5.2
-	google.golang.org/protobuf v1.26.0
+	google.golang.org/protobuf v1.28.0
 )
