+++ pkg/converter/comments.go
@@ -1,4 +1,4 @@
-package main
+package converter
 
 import (
 	"strconv"
