+++ pkg/converter/comments_test.go
@@ -1,10 +1,10 @@
-package main
+package converter
 
 import (
 	"reflect"
 	"testing"
 
-	"github.com/golang/protobuf/protoc-gen-go/descriptor"
+	descriptor "google.golang.org/protobuf/types/descriptorpb"
 )
 
 func TestParseComments(t *testing.T) {
@@ -16,12 +16,12 @@ func TestParseComments(t *testing.T) {
 		&descriptor.FileDescriptorProto{
 			SourceCodeInfo: &descriptor.SourceCodeInfo{
 				Location: []*descriptor.SourceCodeInfo_Location{
-					&descriptor.SourceCodeInfo_Location{
+					{
 						Path:             []int32{4, 0},
 						LeadingComments:  &leadingComment,
 						TrailingComments: &trailingComment,
 					},
-					&descriptor.SourceCodeInfo_Location{
+					{
 						Path:             []int32{4, 0, 3, 0, 2, 0},
 						LeadingComments:  &subMessageFieldLeadingComment,
 						TrailingComments: nil,
@@ -46,7 +46,7 @@ func TestParseCommentsWithoutComments(t *testing.T) {
 		&descriptor.FileDescriptorProto{
 			SourceCodeInfo: &descriptor.SourceCodeInfo{
 				Location: []*descriptor.SourceCodeInfo_Location{
-					&descriptor.SourceCodeInfo_Location{
+					{
 						Path: []int32{4, 0},
 					},
 				},
@@ -67,7 +67,7 @@ func TestCommentsGet(t *testing.T) {
 		&descriptor.FileDescriptorProto{
 			SourceCodeInfo: &descriptor.SourceCodeInfo{
 				Location: []*descriptor.SourceCodeInfo_Location{
-					&descriptor.SourceCodeInfo_Location{
+					{
 						Path:            []int32{4, 0},
 						LeadingComments: &comment,
 					},
